Move account service class
